chore: properly define passenger_count and rate_code_id as INT

+++ redshift/star_schema_setup.sql
@@ -34,7 +34,7 @@ CREATE TABLE dim_locations (
     SORTKEY(borough);
 
 CREATE TABLE dim_rate_codes (
-    rate_code_id FLOAT NOT NULL PRIMARY KEY,
+    rate_code_id INT NOT NULL PRIMARY KEY,
     rate_code_name VARCHAR
 )
     DISTSTYLE ALL;
@@ -52,11 +52,11 @@ CREATE TABLE fact_trips (
     pickup_time_id INT REFERENCES dim_times(time_id),
     dropoff_date_id INT REFERENCES dim_dates(date_id),
     dropoff_time_id INT REFERENCES dim_times(time_id),
-    passenger_count FLOAT,
+    passenger_count INT,
     trip_distance FLOAT,
     pickup_location_id INT REFERENCES dim_locations(location_id),
     dropoff_location_id INT REFERENCES dim_locations(location_id),
-    rate_code_id FLOAT REFERENCES dim_rate_codes(rate_code_id),
+    rate_code_id INT REFERENCES dim_rate_codes(rate_code_id),
     store_and_forward BOOLEAN,
     payment_type_id INT REFERENCES dim_payment_types(payment_type_id),
     fare_amount FLOAT,
